test: use fixture for isolation=none coverage runner script
